src/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mrp-multi-db-connector"
version = "0.1.0"
description = "Cross-department data integration platform for Phison Electronics"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "cx_Oracle>=8.3.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-ldap>=3.4.0",
    "cryptography>=43.0.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../../tests/backend"]
